--- databases/Procfile
@@ -1,1 +1,0 @@
-web: node server.js
