--- databases/npm-debug.log
@@ -1,45 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ '/usr/local/bin/node', '/usr/local/bin/npm', 'start' ]
-2 info using npm@3.8.6
-3 info using node@v6.0.0
-4 verbose run-script [ 'prestart', 'start', 'poststart' ]
-5 info lifecycle tech-trends-dbs@1.0.0~prestart: tech-trends-dbs@1.0.0
-6 silly lifecycle tech-trends-dbs@1.0.0~prestart: no script for prestart, continuing
-7 info lifecycle tech-trends-dbs@1.0.0~start: tech-trends-dbs@1.0.0
-8 verbose lifecycle tech-trends-dbs@1.0.0~start: unsafe-perm in lifecycle true
-9 verbose lifecycle tech-trends-dbs@1.0.0~start: PATH: /usr/local/lib/node_modules/npm/bin/node-gyp-bin:/Users/administrator/Desktop/hera/TechTrends/databases/node_modules/.bin:/usr/local/bin:/usr/local/bin:/usr/local/bin:/usr/local/bin:/usr/local/bin:/usr/local/bin:/usr/local/bin:/usr/local/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin
-10 verbose lifecycle tech-trends-dbs@1.0.0~start: CWD: /Users/administrator/Desktop/hera/TechTrends/databases
-11 silly lifecycle tech-trends-dbs@1.0.0~start: Args: [ '-c', 'node server.js' ]
-12 silly lifecycle tech-trends-dbs@1.0.0~start: Returned: code: 1  signal: null
-13 info lifecycle tech-trends-dbs@1.0.0~start: Failed to exec start script
-14 verbose stack Error: tech-trends-dbs@1.0.0 start: `node server.js`
-14 verbose stack Exit status 1
-14 verbose stack     at EventEmitter.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/lifecycle.js:239:16)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at EventEmitter.emit (events.js:191:7)
-14 verbose stack     at ChildProcess.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/spawn.js:24:14)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at ChildProcess.emit (events.js:191:7)
-14 verbose stack     at maybeClose (internal/child_process.js:850:16)
-14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:215:5)
-15 verbose pkgid tech-trends-dbs@1.0.0
-16 verbose cwd /Users/administrator/Desktop/hera/TechTrends/databases
-17 error Darwin 15.5.0
-18 error argv "/usr/local/bin/node" "/usr/local/bin/npm" "start"
-19 error node v6.0.0
-20 error npm  v3.8.6
-21 error code ELIFECYCLE
-22 error tech-trends-dbs@1.0.0 start: `node server.js`
-22 error Exit status 1
-23 error Failed at the tech-trends-dbs@1.0.0 start script 'node server.js'.
-23 error Make sure you have the latest version of node.js and npm installed.
-23 error If you do, this is most likely a problem with the tech-trends-dbs package,
-23 error not with npm itself.
-23 error Tell the author that this fails on your system:
-23 error     node server.js
-23 error You can get information on how to open an issue for this project with:
-23 error     npm bugs tech-trends-dbs
-23 error Or if that isn't available, you can get their info via:
-23 error     npm owner ls tech-trends-dbs
-23 error There is likely additional logging output above.
-24 verbose exit [ 1, true ]
