--- databases/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "tech-trends-dbs",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon --debug server.js"
-  },
-  "author": "Justin Roche <jproche5@gmail.com>",
-  "license": "ISC",
-  "dependencies": {
-    "body-parser": "^1.15.2",
-    "express": "^4.14.0",
-    "mongo": "^0.1.0",
-    "mongoose": "^4.6.4"
-  }
-}
