+++ deprecated/server.js
@@ -1,26 +1,16 @@
-
-
-const postingsHelpers = require('./database/postingsHelpers.js');
+const POSTINGSURI = 'mongodb://hera:hackreactor19@ds063406.mlab.com:63406/rawpostings'//"mongodb://localhost/postings";
+const postingsHelpers = require('./postingsHelpers.js');
 const bodyParser = require('body-parser');
 var app = require('express')();
 var mongoose = require("mongoose");
 
 //mongoose is connected anywhere, it's connection is referenced whenever it is required
-
-
-const POSTINGSURI = process.env['dev']? 
-      "mongodb://localhost/postings" : 
-      "mongodb://hera:hackreactor19@ds063406.mlab.com:63406/rawpostings";
-
-console.log('URI=========',POSTINGSURI);
 mongoose.connect(POSTINGSURI);
 
-console.log(process.env);
+
 //middleware
 app.use(bodyParser.json());
 
-//change
-//another change
 
 //----------routes for the raw postings database---------------
 //-------------------------------------------------------------
@@ -28,24 +18,17 @@ app.use(bodyParser.json());
 app.post('/raw-postings',function(req,res){
   postingsHelpers.addNewPosting(req.body, (newPosting) => {
     console.log("added new posting", newPosting);
-    res.status(202).send(newPosting);
+    res.status(301).send(newPosting);
   });
+
 });
 
 
 app.get('/raw-postings',function(req,res){
   postingsHelpers.getPostings(req.query.date, function(results){
-    res.status(202).send(results);
+    res.status(200).send(results);
   });
 });
-
-app.delete('/raw-postings',function(req,res){
-  postingsHelpers.deletePostings(req.query.date, function(result){
-    console.log('callback called');
-    res.status(204).send(result);
-  });
-  
-})
 
 //----------routes for the analyzed database-------------------
 //-------------------------------------------------------------
@@ -54,5 +37,3 @@ app.delete('/raw-postings',function(req,res){
 //--end routes--
 //server listen
 app.listen(process.env.PORT || 8000);
-
-module.exports = app;
