+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "NODE_ENV=dev nodemon --debug server.js",
     "start": "nodemon --debug server.js"
   },
   "author": "Justin Roche <jproche5@gmail.com>",
@@ -14,5 +14,9 @@
     "express": "^4.14.0",
     "mongo": "^0.1.0",
     "mongoose": "^4.6.4"
+  },
+  "devDependencies": {
+    "chai": "^3.5.0",
+    "supertest": "^2.0.1"
   }
 }
